refactor(util): 패키지 변경 완료
- git 조작 실수 이슈로, 패키지가 잘못 선언되어있었던 것 수정
